Move most SymPy functions to their own file
All these are standalone implementations of some functions and they
don't depend on anything else, so we better have them under the
`_sympy/` folder on their own

[ghstack-poisoned]
