+++ torch/_inductor/bounds.py
@@ -4,8 +4,9 @@ from typing import Dict, Optional
 import sympy
 
 import torch
+from torch.utils._sympy.functions import FloorDiv, ModularIndexing
 from torch.utils._sympy.value_ranges import ValueRangeAnalysis, ValueRanges
-from .ir import FloorDiv, InterpreterShim, LoopBody, ModularIndexing
+from .ir import InterpreterShim, LoopBody
 from .utils import cache_on_self, dominated_nodes, sympy_subs
 from .virtualized import V
