+++ torch/_inductor/codegen/triton.py
@@ -14,6 +14,7 @@ import torch
 
 import torch._logging
 from torch._prims_common import is_integer_dtype
+from torch.utils._sympy.functions import FloorDiv, ModularIndexing
 from ..._dynamo.utils import counters
 from .. import config, ir, scheduler
 from ..codecache import code_hash, get_path
@@ -536,11 +537,9 @@ class IterationRangesRoot(IterationRanges):
         Lookup a given RangeTreeEntry, creating it if needed
         """
         if V.graph.sizevars.statically_known_equals(divisor * length, self.numel):
-            expr = ir.FloorDiv(sympy_symbol(f"{self.prefix}index"), divisor)
+            expr = FloorDiv(sympy_symbol(f"{self.prefix}index"), divisor)
         else:
-            expr = ir.ModularIndexing(
-                sympy_symbol(f"{self.prefix}index"), divisor, length
-            )
+            expr = ModularIndexing(sympy_symbol(f"{self.prefix}index"), divisor, length)
 
         if expr not in self.nodes:
             node = IterationRangesEntry(
@@ -585,12 +584,12 @@ class IterationRangesRoot(IterationRanges):
         for node in nodes:
             if not V.graph.sizevars.statically_known_equals(node.divisor, divisor):
                 # fill in unused index var
-                add(self.lookup(divisor, ir.FloorDiv(node.divisor, divisor)))
+                add(self.lookup(divisor, FloorDiv(node.divisor, divisor)))
                 divisor = node.divisor
             add(node)
         if not V.graph.sizevars.statically_known_equals(self.numel, divisor):
             # fill in unused index var
-            add(self.lookup(divisor, ir.FloorDiv(self.numel, divisor)))
+            add(self.lookup(divisor, FloorDiv(self.numel, divisor)))
 
         return list(reversed(index_vars)), list(reversed(sizes))
 
@@ -683,7 +682,7 @@ class IterationRangesEntry(IterationRanges):
         precomputed_args = []
         if isinstance(self.expr, sympy.Symbol):
             return precomputed_args
-        assert isinstance(self.expr, (ir.FloorDiv, ir.ModularIndexing)), type(self.expr)
+        assert isinstance(self.expr, (FloorDiv, ModularIndexing)), type(self.expr)
         for arg in self.expr.args[1:]:
             if not isinstance(arg, (sympy.Integer, sympy.Symbol)):
                 symbols = arg.free_symbols
@@ -842,7 +841,7 @@ class TritonKernel(Kernel):
             if not sv.statically_known_multiple_of(remaining[i], expr):
                 raise CantSplit()
             # guard on the last item out
-            remaining[i] = ir.FloorDiv(remaining[i], expr)
+            remaining[i] = FloorDiv(remaining[i], expr)
             new_ranges[i].append(expr)
             return next(var_count)
 
@@ -875,7 +874,7 @@ class TritonKernel(Kernel):
                     ):
                         raise CantSplit()
                     size1 = remaining[current_group]
-                    size2 = ir.FloorDiv(size, remaining[current_group])
+                    size2 = FloorDiv(size, remaining[current_group])
                     return_getters.append(
                         make_combined(
                             size2,
@@ -2362,7 +2361,7 @@ class TritonScheduling:
                     b0, b1 = ranked_tilings[0]
                 assert V.graph.sizevars.size_hint(a1 - b1) > 0
                 if V.graph.sizevars.statically_known_multiple_of(a1, b1):
-                    tiling = (a0, ir.FloorDiv(a1, b1), b1)
+                    tiling = (a0, FloorDiv(a1, b1), b1)
                     ranked_tilings = [tiling] + ranked_tilings
                     break  # only 1 choice for now
