+++ torch/utils/_sympy/interp.py
@@ -14,6 +14,7 @@ import sympy
 from sympy.logic.boolalg import Boolean as SympyBoolean, BooleanAtom
 
 import torch
+from torch.utils._sympy.functions import CleanDiv, FloorDiv
 
 
 # TODO: Dedupe this with SYMPY_INTERP
@@ -21,8 +22,7 @@ import torch
 
 @functools.lru_cache(None)
 def handlers():
-    from torch._inductor.ir import CleanDiv
-    from torch.fx.experimental.symbolic_shapes import FloorDiv, Pow, TrueDiv
+    from torch.fx.experimental.symbolic_shapes import Pow, TrueDiv
 
     # TODO: CeilDiv, ModularIndexing
